2023 up to 7 (kinda); 2024-16; utils overhaul
It's kind of embarrassing how bad I am at even the first week of 2023.
2024-16 was a beast but brute force >>>> [insert smart graph algo]
Also a MASSIVE overhaul of the utils lib. As usual, this means that there is no backwards-compatibility but fixing old solutions to use this should be relatively easy, since the ergonomics of utils are now WAY WAY WAY better